Changed load_cluster behavior.
Chef::Config[:cluster_path] is now interpreted as an array of directories that
are expected to contain cluster definitions. ClusterChef.load_cookbook will now
traverse the array in order, looking for the requested cluster definition in
each directory, and attempts to load the first one that it finds.

By default, the cluster_path is set to the directory that cluster chef used to look
for cluster definitions.

To point cluster chef at a new set of directories, add something like the following
to your knife.rb file:

cluster_path ['cluster_chef/clusters',
              'infochimps_chef/clusters',
             ].map { |path| File.join("/home/chris/ics/sysadmin", path) }

This behavior is very similar to the way "cookbook_path" works...