Adds some flexibility when naming keys in AWS relative to ~/.ssh - Checks for ssh_identity_dir/<keypair>_rsa - Checks for ssh_identity_dir/<keypair> - Checks for ssh_identity_dir/id_<keypair> - falls back on ssh_identity_dir/<keypair>.pem
cloud.ssh_identity_dir needs to be set to ~/.ssh to work properly, otherwise
strange things happen in ~/.chef/ec2_keys.

This is all to allow for an AWS keypair name in ec2 to be e.g. "staging" and the
associated file be "~/.ssh/id_staging" or "~/.ssh/staging.rsa".