Fix cross-region launches
When I run ironfan-homebase on a us-east host and launch a cluster in us-west, I get a Errno::ECONNREFUSED, although the host is reachable
a few moments later.  This is probably due to some amazon-specific inter-zone routing quirks.  I do not see this issue when I run my knife
commands on a host outside of the cloud.