knife cluster bootstrap sets Chef::Config[:environment] (fixes #148) I did a little dip into the bootstrap code; in order to make bootstrap work if there are multiple environments in the slice, you'll have to identify the subgroups with each environment and coordinate launching an ssh session on each. knife bootstrap prepares a single template file that draws on the Chef::Config[:environment] setting, and uses that as the command to a knife::Ssh session. at that point it's in Net::Ssh's hands. Anything we did that was non-uniform would lose parallel ssh sessions.
Instead, knife checks the slice and, as long as all the machines you're bootstrapping live in the same environment, sets the Chef::Config and carries on. If you have multiple environments it explains and raises an exception.

So you can still have a cluster with multiple environments, you just have to bootstrap them in groups by environment.