fix for 'cluster list' and adding ability to use per-realm Gemfile
This brings in support for setting the Gemfile according to the
cluster or realm that is being operated on. When operating on the
prod_control cluster, for example, Ironfan will attempt to use the
Gemfile.prod gem file in the current directory. (This requires
shelling out again.) If the file does not exist, Ironfan simply
continues in whatever environment it's been launched in.