fix: error-reporting error in cluster_ssh
The error handler in Chef's ssh knife command relies on @action_nodes
being populated by a search for targeted Chef nodes. We override
Chef's configure_session method, however, which populates this field,
because our method for generating a list of hosts to ssh to is
different.

This commit populates the @action_nodes field within Ironfan's
configure_session method.

This commit addresses
https://github.com/infochimps-labs/ironfan/issues/330

Squashed commit of the following:

commit 5dcddf034bf122193dfeb162dcfe6e9e116a36a5
Author: Josh Bronson <josh.bronson@gmail.com>
Date:   Wed Jan 15 16:41:09 2014 -0600

    populating @action_nodes in cluster_ssh

Change-Id: I999f56b06739af4e502f6e1f3629823e21ab2334