Fix: ensure new fog works with ec2
One fix involves how security groups were being cached in ironfan
with a prepended 'vpc-deadbeef:' but lookups are later performed
without the prefix.  The fix is to cache the security group with
both prefixed and non-prefixed keys, to guarantee non-breakage in
other cases.

The second fix is to text matched in an error message about duplicate
security groups that we were matching/ignoring.  The error text
changed somewhere upstream, so we widened the net on how the text
is matched.

Change-Id: I4fb9f831710d77dc987057ec39b00c16d468e607