fix: resolve move appropriately
Some of the Ironfan code relies on cluster definitions remaining fixed
after all DSL definitions are done, which means resolve! should be
called a final time just after evaluating the DSL.

Furthermore, we'd like to support re-opening realm definitions, which
will entail calling resolve! on a realm. This will re-instantiate any
cluster objects pointed to by the realm, so we'll need to re-register
those clusters with Ironfan in order for the registry to continue
working.

Change-Id: I6b1203004c9b6e0ef5108b184e3df4381b079ef5